requirements.txt
requests
pylightning
buttplug
